requirements.txt
streamlit
google-cloud-vision
pillow
plotly
pandas